Refactor RAG config to use Pydantic dataclasses without arbitrary_types_allowed
- Convert stdlib dataclasses to Pydantic dataclasses
- Add ChromaEmbeddingFunctionWrapper for type safety
- Move SupportedProvider to optional_imports module
- Remove custom schema_utils in favor of Pydantic's built-in handling
- Suppress Pydantic v1/v2 mixing warning for ChromaDB Settings
