+++ src/crewai/rag/config/types.py
@@ -1,6 +1,6 @@
 """Type definitions for RAG configuration."""
 
-from typing import TYPE_CHECKING, Annotated, TypeAlias, Literal
+from typing import TYPE_CHECKING, Annotated, TypeAlias
 from pydantic import Field
 
 from crewai.rag.config.constants import DISCRIMINATOR
@@ -14,11 +14,6 @@ else:
         from crewai.rag.config.optional_imports.providers import (
             MissingChromaDBConfig as ChromaDBConfig,
         )
-
-SupportedProvider = Annotated[
-    Literal["chromadb"],
-    "Supported RAG provider types, add providers here as they become available",
-]
 
 SupportedProviderConfig: TypeAlias = ChromaDBConfig
 RagConfigType: TypeAlias = Annotated[
